refactor: 버튼 재사용성을 높이기 위한 수정 (#29 PR 피드백)
